Move txn filtering out of mysqlSink
