+++ cdc/owner_test.go
@@ -71,7 +71,7 @@ func (h *handlerForPrueDMLTest) PullDDL() (resolvedTs uint64, ddl []*model.DDL, 
 	return uint64(math.MaxUint64), nil, nil
 }
 
-func (h *handlerForPrueDMLTest) ExecDDL(context.Context, string, *model.DDL) error {
+func (h *handlerForPrueDMLTest) ExecDDL(context.Context, string, model.Txn) error {
 	panic("unreachable")
 }
 
@@ -190,12 +190,12 @@ func (h *handlerForDDLTest) PullDDL() (resolvedTs uint64, jobs []*model.DDL, err
 	return h.ddlResolvedTs[h.ddlIndex], []*model.DDL{h.ddls[h.ddlIndex]}, nil
 }
 
-func (h *handlerForDDLTest) ExecDDL(ctx context.Context, sinkURI string, ddl *model.DDL) error {
+func (h *handlerForDDLTest) ExecDDL(ctx context.Context, sinkURI string, txn model.Txn) error {
 	h.mu.Lock()
 	defer h.mu.Unlock()
 	h.ddlExpectIndex++
-	h.c.Assert(ddl, check.DeepEquals, h.ddls[h.ddlExpectIndex])
-	h.c.Assert(ddl.Job.BinlogInfo.FinishedTS, check.Equals, h.currentGlobalResolvedTs)
+	h.c.Assert(txn.DDL, check.DeepEquals, h.ddls[h.ddlExpectIndex])
+	h.c.Assert(txn.DDL.Job.BinlogInfo.FinishedTS, check.Equals, h.currentGlobalResolvedTs)
 	return nil
 }
