+++ cdc/processor.go
@@ -586,6 +586,10 @@ func (p *processor) syncResolved(ctx context.Context) error {
 			if err != nil {
 				return errors.Trace(err)
 			}
+			filterBySchemaAndTable(&txn)
+			if len(txn.DMLs) == 0 {
+				continue
+			}
 			pendingTxns = append(pendingTxns, txn)
 			if len(pendingTxns) >= bulkLimit {
 				if err := flush(ctx); err != nil {
