+++ cdc/sink/mysql_test.go
@@ -232,36 +232,3 @@ func (s EmitSuite) TestShouldExecDelete(c *check.C) {
 	c.Assert(err, check.IsNil)
 	c.Assert(mock.ExpectationsWereMet(), check.IsNil)
 }
-
-type FilterSuite struct{}
-
-var _ = check.Suite(&FilterSuite{})
-
-func (s *FilterSuite) TestFilterDMLs(c *check.C) {
-	t := model.Txn{
-		DMLs: []*model.DML{
-			{Database: "INFORMATIOn_SCHEmA"},
-			{Database: "test"},
-			{Database: "test_mysql"},
-			{Database: "mysql"},
-		},
-		Ts: 213,
-	}
-	filterBySchemaAndTable(&t)
-	c.Assert(t.Ts, check.Equals, uint64(213))
-	c.Assert(t.DDL, check.IsNil)
-	c.Assert(t.DMLs, check.HasLen, 2)
-	c.Assert(t.DMLs[0].Database, check.Equals, "test")
-	c.Assert(t.DMLs[1].Database, check.Equals, "test_mysql")
-}
-
-func (s *FilterSuite) TestFilterDDL(c *check.C) {
-	t := model.Txn{
-		DDL: &model.DDL{Database: "performance_schema"},
-		Ts:  10234,
-	}
-	filterBySchemaAndTable(&t)
-	c.Assert(t.Ts, check.Equals, uint64((10234)))
-	c.Assert(t.DMLs, check.HasLen, 0)
-	c.Assert(t.DDL, check.IsNil)
-}
